feat(ConversationEndTime-Logging): Implemented conversation end time logging
